requirements.txt
Django~=4.2.7
loguru~=0.7.2
urllib3~=2.0.7
psycopg2-binary~=2.9.9
djangorestframework~=3.14.0
celery~=5.3.5
django-cors-headers~=4.3.1
pytz~=2023.3.post1
django-debug-toolbar~=4.2.0
django-celery-beat~=2.5.0
celery~=5.3.4
flower~=2.0.1
gunicorn~=21.2.0
django_celery_results~=2.5.1
redis~=5.0.1
sqlalchemy~=2.0.23
drf-spectacular~=0.26.5
drf-spectacular-sidecar~=2023.10.1
djangorestframework-simplejwt~=5.3.1
django-mptt~=0.15.0
